refactor: extract question component
