+++ src/App.js
@@ -1,8 +1,8 @@
 import './App.css';
 import {HashRouter as Router, Route, Routes} from 'react-router-dom'
 import Home from "./pages/Home/Home";
-import Question from "./pages/Questions/Question";
-import Index from "./pages/Overview";
+import Question from "./pages/Questions";
+import Overview from "./pages/Overview";
 
 function App() {
   return (
@@ -12,7 +12,7 @@ function App() {
           <Routes>
             <Route path="/" element={<Home/>}/>
             <Route path="/question/:index" element={<Question/>}/>
-            <Route path="/overview" element={<Index/>}/>
+            <Route path="/overview" element={<Overview/>}/>
           </Routes>
         </main>
       </div>
