+++ src/pages/Questions/QuestionHeader.jsx
@@ -1,0 +1,39 @@
+import React from 'react';
+import PageHeader from "../../components/Header/PageHeader";
+import {useNavigate} from "react-router-dom";
+
+const QuestionHeader = ({
+                        isCN,
+                        toggleLanguage,
+                        toggleFavourite,
+                        isFavourite,
+                      }) => {
+  const navigate = useNavigate();
+
+  const backHome = () => {
+    navigate('/');
+  }
+
+  const rightIcons = [
+    {
+      name: 'language',
+      action: toggleLanguage,
+      active: isCN,
+    }, {
+      name: 'fav',
+      inactiveName: "fav_reg",
+      action: toggleFavourite,
+      active: isFavourite,
+    }
+  ]
+
+  return (
+    <PageHeader
+      pageTitle={isCN ? "问题" : "Question"}
+      handleBack={backHome}
+      rightIcons={rightIcons}
+    />
+  );
+};
+
+export default QuestionHeader;
