+++ src/pages/Questions/QuestionInfo.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import {THEME_COLOR} from '../../common/common';
+
+const QuestionInfo = ({
+                          currQuestion,
+                          currQuestionIndex,
+                          displayedQuestions,
+                        }) => {
+  return (
+    <div className="content-head">
+      <div className="question-type" style={{color: THEME_COLOR}}>
+        <span>{currQuestion.section}</span>
+      </div>
+      <div className="question-num">
+            <span className="question-index" style={{color: THEME_COLOR}}>
+              {currQuestionIndex + 1}
+            </span>
+        /
+        <span className="question-num-item">
+              {displayedQuestions.length}
+            </span>
+      </div>
+    </div>
+  );
+};
+
+export default QuestionInfo;
